requirements.txt
grpcio==1.78.0
grpcio-tools==1.70.0
protobuf>=5.29.0,<6.0.0
litellm